+++ NEWS.md
@@ -1,6 +1,8 @@
 # plumber2 (development version)
 
 * Fix a bug in constructing empty blocks for extending
+* You can now pass in additional arguments to the OpenAPI doc UI with
+  `api_doc_setting()`
 
 # plumber2 0.1.0